+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Services/MenuPages/AbstractExtensionModuleDocumentationMenuPage.php
@@ -11,13 +11,13 @@ use GatoGraphQL\GatoGraphQL\ModuleResolvers\ModuleResolverInterface;
 abstract class AbstractExtensionModuleDocumentationMenuPage extends AbstractModuleDocsMenuPage
 {
     /**
-     * If opening a Recipe doc in the iframe, do not use tabpanels
+     * If opening a Tutorial lesson doc in the iframe, do not use tabpanels
      */
     protected function useTabpanelForContent(): bool
     {
         if (
             $this->getMenuPageHelper()->isDocumentationScreen()
-            && str_contains(App::query(RequestParams::DOC, ''), '../docs/recipes/')
+            && str_contains(App::query(RequestParams::DOC, ''), '../docs/tutorial/')
         ) {
             return false;
         }