+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -13,6 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Improvements
 
 - Added "Request headers" to GraphiQL clients on single public/private endpoint, and custom endpoints
+- Renamed "Recipes" as "Tutorial"
 
 ### Fixed